quic: avoid using closures and other cleanups
PR-URL: https://github.com/nodejs/quic/pull/29
