+++ lib/internal/quic/core.js
@@ -9,17 +9,19 @@ const {
 
 assertCrypto();
 
+const {
+  getAllowUnauthorized,
+  getSocketType,
+  lookup4,
+  lookup6,
+  validateTransportParams,
+  validateQuicClientSessionOptions,
+  validateQuicSocketOptions,
+} = require('internal/quic/util');
 const util = require('util');
-const { debuglog } = util;
-const debug = debuglog('quic');
 const assert = require('internal/assert');
 const EventEmitter = require('events');
 const { Duplex } = require('stream');
-const { isArrayBufferView } = require('internal/util/types');
-const {
-  isIP,
-  isLegalPort,
-} = require('internal/net');
 const {
   createSecureContext: _createSecureContext
 } = require('tls');
@@ -56,7 +58,6 @@ const {
     ERR_INVALID_ARG_TYPE,
     ERR_INVALID_ARG_VALUE,
     ERR_INVALID_CALLBACK,
-    ERR_OUT_OF_RANGE,
     ERR_QUICSESSION_DESTROYED,
     ERR_QUICSOCKET_CLOSING,
     ERR_QUICSOCKET_DESTROYED,
@@ -116,13 +117,11 @@ const DEFAULT_GROUPS = 'P-256:X25519:P-384:P-521';
 
 const emit = EventEmitter.prototype.emit;
 
-// Lazy load dns
-let dns;
-
 const kAddSession = Symbol('kAddSession');
 const kAddStream = Symbol('kAddStream');
 const kContinueBind = Symbol('kContinueBind');
 const kContinueConnect = Symbol('kContinueConnect');
+const kContinueListen = Symbol('kContinueListen');
 const kExtend = Symbol('kExtend');
 const kHandshake = Symbol('kHandshake');
 const kHandshakePost = Symbol('kHandshakePost');
@@ -143,90 +142,6 @@ const kSocketPending = 1;
 const kSocketBound = 2;
 const kSocketClosing = 3;
 const kSocketDestroyed = 4;
-
-function lazyDNS() {
-  if (!dns)
-    dns = require('dns');
-  return dns;
-}
-
-function validateNumberInRange(val, name, range) {
-  if (typeof val !== 'number')
-    throw new ERR_INVALID_ARG_TYPE(name, 'number', val);
-  if (val < 0)
-    throw new ERR_OUT_OF_RANGE(name, range, val);
-}
-
-function validateTransportParams(params) {
-  const {
-    maxStreamDataBidiLocal =
-    IDX_QUIC_SESSION_MAX_STREAM_DATA_BIDI_LOCAL_DEFAULT,
-    maxStreamDataBidiRemote =
-    IDX_QUIC_SESSION_MAX_STREAM_DATA_BIDI_REMOTE_DEFAULT,
-    maxStreamDataUni =
-    IDX_QUIC_SESSION_MAX_STREAM_DATA_UNI_DEFAULT,
-    maxData =
-    IDX_QUIC_SESSION_MAX_DATA_DEFAULT,
-    maxStreamsBidi =
-    IDX_QUIC_SESSION_MAX_STREAMS_BIDI_DEFAULT,
-    maxStreamsUni =
-    IDX_QUIC_SESSION_MAX_STREAMS_UNI_DEFAULT,
-    idleTimeout =
-    IDX_QUIC_SESSION_IDLE_TIMEOUT_DEFAULT,
-    maxPacketSize =
-    IDX_QUIC_SESSION_MAX_PACKET_SIZE_DEFAULT,
-    maxAckDelay = NGTCP2_DEFAULT_MAX_ACK_DELAY,
-    preferredAddress,
-  } = { ...params };
-  validateNumberInRange(
-    maxStreamDataBidiLocal,
-    'options.maxStreamDataBidiLocal',
-    '>=0');
-  validateNumberInRange(
-    maxStreamDataBidiRemote,
-    'options.maxStreamDataBidiRemote',
-    '>=0');
-  validateNumberInRange(
-    maxStreamDataUni,
-    'options.maxStreamDataUni',
-    '>=0');
-  validateNumberInRange(
-    maxData,
-    'options.maxData',
-    '>=0');
-  validateNumberInRange(
-    maxStreamsBidi,
-    'options.maxStreamsBidi',
-    '>=0');
-  validateNumberInRange(
-    maxStreamsUni,
-    'options.maxStreamsUni',
-    '>=0');
-  validateNumberInRange(
-    idleTimeout,
-    'options.idleTimeout',
-    '>=0');
-  validateNumberInRange(
-    maxPacketSize,
-    'options.maxPacketSize',
-    '>=0');
-  validateNumberInRange(
-    maxAckDelay,
-    'options.maxAckDelay',
-    '>=0');
-  return {
-    maxStreamDataBidiLocal,
-    maxStreamDataBidiRemote,
-    maxStreamDataUni,
-    maxData,
-    maxStreamsBidi,
-    maxStreamsUni,
-    idleTimeout,
-    maxPacketSize,
-    maxAckDelay,
-    preferredAddress,
-  };
-}
 
 function setConfigField(val, index) {
   if (typeof val === 'number') {
@@ -376,31 +291,6 @@ setCallbacks({
   onStreamError,
 });
 
-function getSocketType(type) {
-  switch (type) {
-    case 'udp4': return AF_INET;
-    case 'udp6': return AF_INET6;
-  }
-  throw new ERR_INVALID_ARG_VALUE('options.type', type);
-}
-
-function validateBindOptions(port, address) {
-  if (port != null && typeof port !== 'number')
-    throw new ERR_INVALID_ARG_TYPE('options.port', 'number', port);
-  if (address != null && typeof address !== 'string')
-    throw new ERR_INVALID_ARG_TYPE('options.address', 'string', address);
-}
-
-function lookup4(address, callback) {
-  const { lookup } = lazyDNS();
-  lookup(address || '127.0.0.1', 4, callback);
-}
-
-function lookup6(address, callback) {
-  const { lookup } = lazyDNS();
-  lookup(address || '::1', 6, callback);
-}
-
 function afterLookup(callback, err, ip) {
   if (err) {
     this.destroy(err);
@@ -417,36 +307,41 @@ function connectAfterLookup(type, err, ip) {
   this[kContinueConnect](type, ip);
 }
 
-function validateQuicSocketOptions(options) {
-  const {
-    address,
-    client,
-    ipv6Only = false,
-    lookup,
-    port,
-    reuseAddr = false,
-    server,
-    type = 'udp4'
-  } = { ...options };
-  validateBindOptions(port, address);
-  if (typeof type !== 'string')
-    throw new ERR_INVALID_ARG_TYPE('options.type', 'string', type);
-  if (lookup && typeof lookup !== 'function')
-    throw new ERR_INVALID_ARG_TYPE('options.lookup', 'Function', lookup);
-  if (typeof ipv6Only !== 'boolean')
-    throw new ERR_INVALID_ARG_TYPE('options.ipv6Only', 'boolean', ipv6Only);
-  if (typeof reuseAddr !== 'boolean')
-    throw new ERR_INVALID_ARG_TYPE('options.reuseAddr', 'boolean', reuseAddr);
-  return {
-    address,
-    client,
-    ipv6Only,
-    lookup,
-    port,
-    reuseAddr,
-    server,
-    type: getSocketType(type)
-  };
+function afterPreferredAddressLookup(transportParams, port, type, err, ip) {
+  if (err) {
+    this.destroy(err);
+    return;
+  }
+  this[kContinueListen](transportParams, { ip, port, type });
+}
+
+function continueListen(transportParams, lookup) {
+  const { preferredAddress } = transportParams;
+
+  if (preferredAddress && typeof preferredAddress === 'object') {
+    const {
+      address,
+      port,
+      type = 'udp4',
+    } = { ...preferredAddress };
+    const typeVal = getSocketType(type);
+    // If preferred address is set, we need to perform a lookup on it
+    // to get the IP address. Only after that lookup completes can we
+    // continue with the listen operation, passing in the resolved
+    // preferred address.
+    lookup(
+      address || (typeVal === AF_INET6 ? '::' : '0.0.0.0'),
+      afterPreferredAddressLookup.bind(this, transportParams, port, typeVal));
+    return;
+  }
+  // If preferred address is not set, we can skip directly to the listen
+  this[kContinueListen](transportParams);
+};
+
+function connectAfterBind(session, lookup, address, type) {
+  lookup(
+    address || (type === AF_INET6 ? '::' : '0.0.0.0'),
+    connectAfterLookup.bind(session, type));
 }
 
 // QuicSocket wraps a UDP socket plus the associated TLS context and QUIC
@@ -564,6 +459,22 @@ class QuicSocket extends EventEmitter {
     this[kHandle].receiveStop();
   }
 
+  [kContinueListen](transportParams, preferredAddress) {
+    const {
+      address,
+      port,
+      type = AF_INET,
+    } = { ...preferredAddress };
+    setTransportParams(transportParams);
+    this[kHandle].listen(
+      this.#serverSecureContext.context,
+      address,
+      type,
+      port);
+    process.nextTick(emit.bind(this, 'listening'));
+  };
+
+
   // Begin listening for server connections
   listen(options, callback) {
     if (this.#serverListening)
@@ -601,53 +512,11 @@ class QuicSocket extends EventEmitter {
     this.#serverSecureContext = createSecureContext(options, initSecureContext);
     this.#serverListening = true;
 
-    const transportParams = validateTransportParams(options);
-
-    const doListen = () => {
-      const { preferredAddress } = transportParams;
-
-      const doFinalListen = (preferredAddress) => {
-        const {
-          address,
-          port,
-          type = AF_INET,
-        } = { ...preferredAddress };
-        setTransportParams(transportParams);
-        this[kHandle].listen(
-          this.#serverSecureContext.context,
-          address,
-          type,
-          port);
-        process.nextTick(emit.bind(this, 'listening'));
-      };
-
-      if (preferredAddress && typeof preferredAddress === 'object') {
-        const {
-          address,
-          port,
-          type = 'udp4',
-        } = { ...preferredAddress };
-        const typeVal = getSocketType(type);
-        // If preferred address is set, we need to perform a lookup on it
-        // to get the IP address. Only after that lookup completes can we
-        // continue with the listen operation, passing in the resolved
-        // preferred address.
-        this.#lookup(
-          address || (typeVal === AF_INET6 ? '::' : '0.0.0.0'),
-          (err, address) => {
-            if (err) {
-              // TODO(@jasnell): More descriptive error
-              this.destroy(err);
-              return;
-            }
-            doFinalListen({ address, port, type: typeVal });
-          });
-        return;
-      }
-      // If preferred address is not set, we can skip directly
-      // to the listen
-      doFinalListen();
-    };
+    const doListen =
+      continueListen.bind(
+        this,
+        validateTransportParams(options),
+        this.#lookup);
 
     // If the QuicSocket is already bound, we'll begin listening
     // immediately. If we're still pending, however, wait until
@@ -686,12 +555,13 @@ class QuicSocket extends EventEmitter {
     if (typeof callback === 'function')
       session.on('ready', callback);
 
-    const typeVal = getSocketType(type);
-    this[kMaybeBind](() => {
-      this.#lookup(
-        address || (typeVal === AF_INET6 ? '::' : '0.0.0.0'),
-        connectAfterLookup.bind(session, typeVal));
-    });
+    this[kMaybeBind](
+      connectAfterBind.bind(
+        this,
+        session,
+        this.#lookup,
+        address,
+        getSocketType(type)));
 
     return session;
   }
@@ -979,6 +849,9 @@ class QuicSession extends EventEmitter {
   }
 
   [kHandshakePost]() {
+    // Non-op for the default case. QuicClientSession
+    // overrides this with some client-side specific
+    // checks
     return true;
   }
 
@@ -1187,7 +1060,32 @@ class QuicServerSession extends QuicSession {
   }
 }
 
-let warnOnAllowUnauthorized = true;
+function setSocketAfterBind(socket, callback) {
+  if (!this.readyToMigrate) {
+    callback(new ERR_QUICSESSION_UNABLE_TO_MIGRATE());
+    return;
+  }
+
+  if (socket.destroyed) {
+    callback(new ERR_QUICSOCKET_DESTROYED('setSocket'));
+    return;
+  }
+
+  const err = this[kHandle].setSocket(socket[kHandle]);
+  if (err !== 0) {
+    callback(new ERR_QUICCLIENTSESSION_FAILED_SETSOCKET(err));
+    return;
+  }
+
+  if (this.socket) {
+    this.socket[kRemoveSession](this);
+    this[kSetSocket](undefined);
+  }
+  socket[kAddSession](this);
+  this[kSetSocket](socket);
+
+  callback();
+}
 
 class QuicClientSession extends QuicSession {
   #handleReady = false;
@@ -1201,64 +1099,20 @@ class QuicClientSession extends QuicSession {
   #remoteTransportParams = undefined;
 
   constructor(socket, options) {
-    const allowUnauthorized = process.env.NODE_TLS_REJECT_UNAUTHORIZED === '0';
-
-    if (allowUnauthorized && warnOnAllowUnauthorized) {
-      warnOnAllowUnauthorized = false;
-      process.emitWarning(
-        'Setting the NODE_TLS_REJECT_UNAUTHORIZED ' +
-        'environment variable to \'0\' makes TLS connections ' +
-        'and HTTPS requests insecure by disabling ' +
-        'certificate verification.');
-    }
-
     const sc_options = {
       secureProtocol: 'TLSv1_3_client_method',
-      rejectUnauthorized: !allowUnauthorized,
+      rejectUnauthorized: !getAllowUnauthorized(),
       ...options
     };
 
     const {
-      address,
-      servername = address,
-      port = 0,
-      ipv6Only = false,
-      minDHSize = 1024,
+      servername,
+      port,
+      ipv6Only,
+      minDHSize,
       remoteTransportParams,
       sessionTicket,
-    } = options;
-
-    if (typeof minDHSize !== 'number')
-      throw new ERR_INVALID_ARG_TYPE(
-        'options.minDHSize', 'number', minDHSize);
-
-    if (!isLegalPort(port)) {
-      throw new ERR_INVALID_ARG_VALUE(
-        'options.port', port,
-        'is not a valid IP port');
-    }
-
-    if (servername && typeof servername !== 'string') {
-      throw new ERR_INVALID_ARG_TYPE(
-        'options.servername', 'string', servername);
-    }
-    if (isIP(servername)) {
-      throw new ERR_INVALID_ARG_VALUE(
-        'options.servername', servername, 'cannot be an IP address');
-    }
-
-    if (remoteTransportParams && !isArrayBufferView(remoteTransportParams)) {
-      throw new ERR_INVALID_ARG_TYPE(
-        'options.remoteTransportParams',
-        ['Buffer', 'TypedArray', 'DataView'],
-        remoteTransportParams);
-    }
-    if (sessionTicket && !isArrayBufferView(sessionTicket)) {
-      throw new ERR_INVALID_ARG_TYPE(
-        'options.sessionTicket',
-        ['Buffer', 'TypedArray', 'DataView'],
-        sessionTicket);
-    }
+    } = validateQuicClientSessionOptions(options);
 
     super(socket, servername);
     this.#transportParams = validateTransportParams(options);
@@ -1272,9 +1126,9 @@ class QuicClientSession extends QuicSession {
   }
 
   [kHandshakePost]() {
-    const ekeyinfo = this.ephemeralKeyInfo;
-    if (ekeyinfo.type === 'DH' && ekeyinfo.size < this.#minDHSize) {
-      session.destroy(new ERR_TLS_DH_PARAM_SIZE(ekeyinfo.size));
+    const { type, size } = this.ephemeralKeyInfo;
+    if (type === 'DH' && size < this.#minDHSize) {
+      session.destroy(new ERR_TLS_DH_PARAM_SIZE(size));
       return false;
     }
     return true;
@@ -1359,32 +1213,7 @@ class QuicClientSession extends QuicSession {
     if (typeof callback !== 'function')
       throw new ERR_INVALID_CALLBACK();
 
-    socket[kMaybeBind](() => {
-      if (!this.readyToMigrate) {
-        callback(new ERR_QUICSESSION_UNABLE_TO_MIGRATE());
-        return;
-      }
-
-      if (socket.destroyed) {
-        callback(new ERR_QUICSOCKET_DESTROYED('setSocket'));
-        return;
-      }
-
-      const err = this[kHandle].setSocket(socket[kHandle]);
-      if (err !== 0) {
-        callback(new ERR_QUICCLIENTSESSION_FAILED_SETSOCKET(err));
-        return;
-      }
-
-      if (this.socket) {
-        this.socket[kRemoveSession](this);
-        this[kSetSocket](undefined);
-      }
-      socket[kAddSession](this);
-      this[kSetSocket](socket);
-
-      callback();
-    });
+    socket[kMaybeBind](setSocketAfterBind.bind(this, socket, callback));
   }
 }
